Send an update to the UI for every 1% of progress in the sim

+++ src/cpp/spell.h
@@ -60,7 +60,7 @@ struct LifeTap : public Spell
 {
     LifeTap(Player* player);
     bool ready();
-    double manaGain();
+    int manaGain();
     void cast();
 };
